requirements.txt
fastapi
uvicorn
psd-tools
Pillow
